refactored local funcs in meterdef controller, added provider for Prometheus API
